major communication protocol architecure changes
