+++ src/lib/interfaces/IOpenID4VCI.ts
@@ -1,0 +1,8 @@
+import { CredentialConfigurationSupported } from "../schemas/CredentialConfigurationSupportedSchema";
+
+export interface IOpenID4VCI {
+	handleCredentialOffer(credentialOfferURL: string): Promise<{ credentialIssuer: string, selectedCredentialConfigurationId: string; issuer_state?: string }>;
+	getAvailableCredentialConfigurations(credentialIssuerIdentifier: string): Promise<Record<string, CredentialConfigurationSupported>>;
+	generateAuthorizationRequest(credentialIssuerIdentifier: string, credentialConfigurationId: string, issuer_state?: string): Promise<{ url?: string }>;
+	handleAuthorizationResponse(url: string, dpopNonceHeader?: string): Promise<void>;
+}
