--- src/lib/interfaces/IOpenID4VCIClient.ts
@@ -1,8 +1,0 @@
-import { CredentialConfigurationSupported } from "../schemas/CredentialConfigurationSupportedSchema";
-
-export interface IOpenID4VCIClient {
-	handleCredentialOffer(credentialOfferURL: string, userHandleB64u: string): Promise<{ credentialIssuer: string, selectedCredentialConfigurationId: string; issuer_state?: string }>;
-	getAvailableCredentialConfigurations(): Promise<Record<string, CredentialConfigurationSupported>>;
-	generateAuthorizationRequest(credentialConfigurationId: string, userHandleB64u: string, issuer_state?: string): Promise<{ url?: string }>;
-	handleAuthorizationResponse(url: string, userHandleB64U: string, dpopNonceHeader?: string): Promise<void>;
-}
