+++ src/lib/interfaces/IOpenID4VCIClientStateRepository.ts
@@ -1,9 +1,9 @@
 import { OpenID4VCIClientState } from "../types/OpenID4VCIClientState";
 
 export interface IOpenID4VCIClientStateRepository {
-	getByStateAndUserHandle(state: string, userHandleB64U: string): Promise<OpenID4VCIClientState | null>;
-	getByCredentialIssuerIdentifierAndCredentialConfigurationIdAndUserHandle(credentialIssuerIdentifier: string, credentialConfigurationId: string, userHandleB64U: string): Promise<OpenID4VCIClientState | null>;
+	getByStateAndUserHandle(state: string): Promise<OpenID4VCIClientState | null>;
+	getByCredentialIssuerIdentifierAndCredentialConfigurationIdAndUserHandle(credentialIssuerIdentifier: string, credentialConfigurationId: string): Promise<OpenID4VCIClientState | null>;
 	create(s: OpenID4VCIClientState): Promise<void>;
-	updateState(s: OpenID4VCIClientState, userHandleB64U: string): Promise<void>;
-	cleanupExpired(userHandleB64U: string): Promise<void>;
+	updateState(s: OpenID4VCIClientState): Promise<void>;
+	cleanupExpired(): Promise<void>;
 }
