+++ src/lib/interfaces/IOpenID4VP.ts
@@ -1,4 +1,4 @@
-export interface IOpenID4VPRelyingParty {
+export interface IOpenID4VP {
 	handleAuthorizationRequest(url: string): Promise<{ conformantCredentialsMap: Map<string, string[]>, verifierDomainName: string } | { err: HandleAuthorizationRequestError }>;
 	promptForCredentialSelection(conformantCredentialsMap: {[x: string]: string[]}, verifierDomainName: string): Promise<Map<string, string>>;
 	sendAuthorizationResponse(selectionMap: Map<string, string>): Promise<{ url?: string } | { presentation_during_issuance_session: string }>;
@@ -10,4 +10,5 @@ export enum HandleAuthorizationRequestError {
 	MISSING_PRESENTATION_DEFINITION_URI,
 	NONTRUSTED_VERIFIER,
 	INVALID_RESPONSE_MODE,
+	OLD_STATE,
 }
